backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.2
sqlalchemy==2.0.23
aiofiles==23.2.1
python-multipart==0.0.6
sse-starlette==1.6.5
psycopg2-binary==2.9.7 